+++ website/content/commands/config/delete.mdx
@@ -56,6 +56,8 @@ Usage: `consul config delete [options]`
   `proxy-defaults` config entry must be `global`, and the name of the `mesh`
   config entry must be `mesh`.
 
+- `-filename` - Specifies the file of the config entry to read.
+
 - `-cas` - Perform a Check-And-Set operation. Specifying this value also
   requires the -modify-index flag to be set. The default value is false.
 
@@ -67,3 +69,5 @@ Usage: `consul config delete [options]`
     $ consul config delete -kind service-defaults -name web
 
     $ consul config delete -kind service-defaults -name web -cas -modify-index 26
+
+    $ consul config delete -filename service-defaults-web.hcl
